VIITE-1716 Fix type conversion

+++ digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/dao/RoadwayDAO.scala
@@ -738,7 +738,7 @@ class RoadwayDAO extends BaseDAO {
       """.as[Long].list
   }
 
-  def getValidBetweenRoadNumbers(roadNumbers: (Int, Int)): List[Long] = {
+  def getValidBetweenRoadNumbers(roadNumbers: (Long, Long)): List[Long] = {
     sql"""
        select distinct road_number
               from ROADWAY
